refactor: ExceptionHandler 위치 이동, null 검사 추가
