+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -10,7 +10,7 @@ import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import org.apache.coyote.http11.Http11Processor;
 import org.apache.coyote.http11.RequestMapping;
-import org.apache.coyote.http11.controller.ExceptionHandler;
+import org.apache.coyote.http11.ExceptionHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
