+++ tomcat/src/main/java/org/apache/coyote/http11/ExceptionHandler.java
@@ -1,10 +1,11 @@
-package org.apache.coyote.http11.controller;
+package org.apache.coyote.http11;
 
 import java.util.List;
 import org.apache.coyote.common.HttpRequest;
 import org.apache.coyote.common.HttpResponse;
 import org.apache.coyote.exception.MethodNotAllowedException;
 import org.apache.coyote.exception.ResourceNotFoundException;
+import org.apache.coyote.http11.controller.Controller;
 
 public class ExceptionHandler {
 
@@ -12,8 +13,19 @@ public class ExceptionHandler {
     private final Controller notFoundController;
 
     private ExceptionHandler(Controller methodNotAllowedController, Controller notFoundController) {
+        validate(methodNotAllowedController, notFoundController);
+
         this.methodNotAllowedController = methodNotAllowedController;
         this.notFoundController = notFoundController;
+    }
+
+    private void validate(Controller methodNotAllowedController, Controller notFoundController) {
+        if (methodNotAllowedController == null) {
+            throw new IllegalArgumentException("MethodNotAllowedController는 null이 될 수 없습니다.");
+        }
+        if (notFoundController == null) {
+            throw new IllegalArgumentException("NotFoundController는 null이 될 수 없습니다.");
+        }
     }
 
     public static ExceptionHandlerBuilder builder() {
